Move values func into helpers js.
